[Fix] Temporarily hack main entry, so it's compatible with other resolvers
This package can't be included right now with webpack (Jest also has a similar issue).

While it's a bug in webpack, this package is not usable by any web app right now that bundles with webpack.
https://github.com/webpack/enhanced-resolve/issues/123

By switching the "main" entry in package.json, this package will be compatible with all resolvers